+++ UM2N/generator/__init__.py
@@ -3,6 +3,6 @@ from .equation_solver import *  # noqa
 from .rand_source_generator import *  # noqa
 from .equation_generator import *  # noqa
 
-from .meshgen import *  # noqa
+from .unstructured_mesh import *  # noqa
 from .burgers_solver import *  # noqa
 from .swirl_solver import *  # noqa
